实侧话网格
