+++ pre-eval/evaluation_eqdis/config
@@ -10,7 +10,7 @@ PreDuration=720
 PreTimeLimit=0/720,0/360,360/720
 NeighborhoodRange=0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15
 EvaluationMethod=N1,N2,N3,N4,POD,BIAS,FAR,FOM,TS,ETS,PC,FSS
-lonBegin=1
-lonEnd=2
-latBegin=3
-latEnd=4
+lonBegin=106
+lonEnd=200
+latBegin=24
+latEnd=200
